Session stats recording

+++ electron/cronjobs/check3x.ts
@@ -5,6 +5,19 @@ const check3x: Cronjob = {
   onTick: function () {
     if (!global.db) return;
 
+    const exists: boolean = db
+      .prepare(
+        `
+        SELECT 1
+        FROM sqlite_master 
+        WHERE type='table' 
+        AND name='player_logs';
+      `
+      )
+      .get();
+
+    if (!exists) return;
+
     global.db
       .prepare(
         `DELETE FROM player_logs WHERE ${Math.floor(
